+++ src/cryptonote_core/checkpoints_create.h
@@ -29,26 +29,20 @@
 // Parts of this file are originally copyright (c) 2012-2013 The Cryptonote developers
 
 #pragma once
-#include <map>
-#include <vector>
-#include "cryptonote_basic_impl.h"
 
+#include "checkpoints.h"
+#include "misc_log_ex.h"
+
+#define ADD_CHECKPOINT(h, hash)  CHECK_AND_ASSERT(checkpoints.add_checkpoint(h,  hash), false);
+#define JSON_HASH_FILE_NAME "checkpoints.json"
 
 namespace cryptonote
 {
-  class checkpoints
-  {
-  public:
-    checkpoints();
-    bool add_checkpoint(uint64_t height, const std::string& hash_str);
-    bool is_in_checkpoint_zone(uint64_t height) const;
-    bool check_block(uint64_t height, const crypto::hash& h) const;
-    bool check_block(uint64_t height, const crypto::hash& h, bool& is_a_checkpoint) const;
-    bool is_alternative_block_allowed(uint64_t blockchain_height, uint64_t block_height) const;
-    uint64_t get_max_height() const;
-    const std::map<uint64_t, crypto::hash>& get_points() const;
-    bool check_for_conflicts(const checkpoints& other) const;
-  private:
-    std::map<uint64_t, crypto::hash> m_points;
-  };
-}
+
+  bool create_checkpoints(cryptonote::checkpoints& checkpoints);
+
+  bool load_checkpoints_from_json(cryptonote::checkpoints& checkpoints, std::string json_hashfile_fullpath);
+  bool load_checkpoints_from_dns(cryptonote::checkpoints& checkpoints, bool testnet = false);
+  bool load_new_checkpoints(cryptonote::checkpoints& checkpoints, std::string json_hashfile_fullpath);
+
+}  // namespace cryptonote
