+++ src/daemon/core.h
@@ -28,6 +28,7 @@
 
 #pragma once
 
+#include "cryptonote_core/checkpoints_create.h"
 #include "cryptonote_core/cryptonote_core.h"
 #include "cryptonote_protocol/cryptonote_protocol_handler.h"
 #include "misc_log_ex.h"
